The Hackathon
Credit Suisse has multiple departments; they are named as Department1, Department2 and Department3 respectively. We want to create a group of employees from these three departments so that they can participate in a hackathon. But there are some rules which we need to keep in mind while forming the groups.

• There is a limit on the group size. A group can have a specified minimum and maximum number of employees.

• The maximum number of employees from each department (Department1, Department2 and Department3) that can be part of a single group is also known.

Our aim is to process some requests in the order in which they are given. Each of the request contains a pair of employees who want to be in the same group. For each such request, the employees are placed in the same group if:

• The size of the group formed does not exceed the given maximum group limit.

• The number of employees in each department does not exceed the maximum for that particular department.

The request is ignored if these conditions are not met.

After all the requests are considered, the organizer discards all groups with less than the minimum required members, and all remaining groups join the contest.

Complete the function theHackathon and print the names of all the employees present in the groups that have the maximum number of members. Print "no groups" if there are no groups. Print these names in lexicographical order, on separate lines. You need to take the employee information and requests from the standard input as described in the input format section.

Constraints:
5 <= n <= 104
5 <= m <= 5x103
2 <= a <= 100
2 <= b <= 200
2 <= f,s,t <= 50
namei is an alphanumeric string with at most 10 characters
departmenti belongs to {1,2,3}
All names are unique
xi, yi are names of existing employees
Input Format:
The first line of input contains seven space-separated integers n,m,a,b,f,s,t.

n -total number of employees
m -number of requests
a and b -minimum and maximum number of employees in a group, respectively.
f, s and t -maximum number of employees from the three departments (Department1, Department2 and Department3) in a group, respectively.
The ith of the next n lines contains namei and departmenti, separated by a space, denoting the name and department of the ith employee, respectively.

The ith of the next m lines describe the ith query and contains two space-separated strings xi, yi - denoting that employees xi and yi want to be placed in the same group.

Output Format:
If there are no groups in the contest, print "no groups". Otherwise, print the names of all the employees contained in all largest groups, printing each name on a new line.
The names must be sorted in increasing lexicographical order according to their ASCII values.
Do not print Debug Statements while submitting the solution.